Consolidate LabelsSet I/O tests and finalize API
- Remove separate load_slp_labels_set and load_ultralytics_labels_set functions
  in favor of unified load_labels_set with format parameter
- Consolidate all LabelsSet tests from separate files into their corresponding
  main test modules (test_main.py, test_slp.py, test_ultralytics.py)
- Update LabelsSet.save() method to support format kwarg for choosing between
  'slp' and 'ultralytics' output formats
- Fix linting issues and apply consistent formatting

This completes the LabelsSet implementation for issue #196.
