+++ sleap_io/__init__.py
@@ -8,8 +8,6 @@ from sleap_io.io.main import (
     load_nwb,
     load_skeleton,
     load_slp,
-    load_slp_labels_set,
-    load_ultralytics_labels_set,
     load_video,
     save_file,
     save_jabs,
@@ -50,8 +48,6 @@ __all__ = [
     "load_nwb",
     "load_skeleton",
     "load_slp",
-    "load_slp_labels_set",
-    "load_ultralytics_labels_set",
     "load_video",
     "save_file",
     "save_jabs",
